chore: rename module directory
