--- src/Application/index.ts
@@ -1,2 +1,0 @@
-export type * from '@/Application/types';
-export * from '@/Application/ApplicationFactory';
