+++ src/application/types.ts
