+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from '@/Application';
+export * from '@/application';
 export * from '@/Config';
 export * from '@/Http';
 export * from '@/Kernel';
